requirements.txt
jupyterlab
Pillow
numpy
matplotlib
ipywidgets
scikit-image